Move lasindex assets to separate module
Because in dependent applications we must use the dagster.load_assets_* functions and the smallest unit is a module.
With this change, we can avoid loading the 200m tiling assets.
